refactor(model): IllustsBean 迁移收尾——清残留命名、gif/相关作品接口 id 放宽 long
- resolveIllustsBean → resolveIllust、launchFromIllustsBeans → launchFromIllusts、mIllustsBean → mIllust
- SlideshowLauncher 两个重载（bean 版 / Illust 版）合并成一个，删掉 bestUrlForV2
- AppApi.relatedIllust / getGifPackage、AppApiSuspend.getGifPackage 的 illust_id 放宽成 long，
  4 处调用点的 (int)/.toInt() 强转删除

Pixel 8 真机冒烟：首页推荐流渲染、卡片收藏 → V3 详情红心同步 → 详情取消收藏 → 列表回灰、
legacy 详情页（useArtworkV3=false）画师行 + 关注按钮、deep link 直达，全程无 FATAL。

+++ app/src/main/java/ceui/lisa/http/AppApiSuspend.kt
@@ -58,7 +58,7 @@ interface AppApiSuspend {
     ): ListUser
 
     @GET("v1/ugoira/metadata")
-    suspend fun getGifPackage(@Query("illust_id") illustId: Int): GifResponse
+    suspend fun getGifPackage(@Query("illust_id") illustId: Long): GifResponse
 
     @GET("v1/illust/detail?filter=for_android")
     suspend fun getIllustByID(@Query("illust_id") illustId: Long): IllustSearchResponse
