+++ app/src/main/java/ceui/pixiv/ui/bulk/UgoiraTask.kt
@@ -125,7 +125,7 @@ suspend fun downloadUgoira(
     val resp: GifResponse = if (cached?.ugoira_metadata != null) {
         cached
     } else {
-        val fetched = Retro.getAppApiSuspend().getGifPackage(illustId.toInt())
+        val fetched = Retro.getAppApiSuspend().getGifPackage(illustId)
         runCatching { Cache.get().saveModel(Params.ILLUST_ID + "_" + illustId, fetched) }
         fetched
     }
